change type of coordinates in enpoint representation to `int_fast32_t`

+++ code/a_star.h
@@ -11,7 +11,7 @@
 // type definition for the type of heuristic fuctions
 typedef uint16_t (*HeuristicFunc)(const Uint16Point p, const Uint16Point goal);
 
-int a_star_cost(const Graph* const g, HeuristicFunc h);
+int16_t a_star_cost(const Graph* const g, HeuristicFunc h);
